Changed ranking type to float

+++ appstore/model/rating_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Column, DateTime, Integer, ForeignKey
+from sqlalchemy import String, Column, DateTime, Integer, ForeignKey, Float
 
 from appstore.schema import rating_schema
 from run import Base
@@ -7,12 +7,10 @@ from run import Base
 class RatingModel(Base):
     __tablename__ = 'rating'
 
-
     id_rating = Column('idrating', Integer, primary_key=True)
-    value = Column(Integer)
+    value = Column(Float)
     comm = Column(String)
     id_app = Column('idapp', ForeignKey('appstore.idapp'))
-
 
     def __init__(self, value, comm, id_app):
         self.value = value
